Updates renderPokemon function to support replacing existing HTML
The renderPokemon function in api.js has been updated to include an optional 'replace' parameter. This correct a bug where the same pokemon was displayed in double when searched by id, by allowing replacing the existing HTML content of the pokemonContainer element with the newly generated HTML, rather than appending it. This change improves the flexibility and control over rendering Pokemon data in the pokedex project.

Fixes #123